A couple of more fixes for more than 4 billion particles: 64 bit arithmetic in loading routines, 64 bit counts in Sorter.

+++ InOutput.C
@@ -184,7 +184,7 @@ void TreePiece::loadTipsy(const std::string& filename,
         int myIndex = CkMyPe();
 	myNumParticles = nTotalParticles / numLoadingPEs;
 	int excess = nTotalParticles % numLoadingPEs;
-	int64_t startParticle = myNumParticles * myIndex;
+	int64_t startParticle = ((int64_t) myNumParticles) * myIndex;
 	if(myIndex < excess) {
 	    myNumParticles++;
 	    startParticle += myIndex;
@@ -999,7 +999,7 @@ void TreePiece::loadNChilada(const std::string& filename,
         int myIndex = CkMyPe();
 	myNumParticles = nTotalParticles / numLoadingPEs;
 	int excess = nTotalParticles % numLoadingPEs;
-	int64_t startParticle = myNumParticles * myIndex;
+	int64_t startParticle = ((int64_t)myNumParticles) * myIndex;
 	if(myIndex < excess) {
 	    myNumParticles++;
 	    startParticle += myIndex;
